+++ js/util.js
@@ -1,3 +1,13 @@
 const isEscapeKey = (evt) => evt.key === 'Escape';
 
-export {isEscapeKey};
+const debounce = (callback, timeoutDelay = 500) => {
+  let timeoutId;
+
+  return (...rest) => {
+    clearTimeout(timeoutId);
+    timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
+  };
+};
+
+
+export {isEscapeKey, debounce};
